refactor(core): extract lifecycle metrics to @principles/core (PRI-52)
Move computeRuleMetrics() and computePrincipleAdherence() from plugin
to core/runtime-v2/internalization/lifecycle-metrics.ts with all 17
magic numbers extracted to named constants. Plugin re-exports from
core for backward compatibility. Architecture regression guards added.
